requirements.txt
# SalesBreachPro Dependencies
# ===========================

# Core Flask Application
Flask==2.3.3
Flask-SQLAlchemy>=3.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3

# Database
SQLAlchemy==2.0.21

# Configuration & Environment
python-dotenv==1.0.0

# HTTP Requests & External APIs
requests==2.31.0

# Data Processing
pandas==2.1.1
python-dateutil==2.8.2
pytz==2023.3

# Email Services
brevo-python==1.2.0
sib-api-v3-sdk==7.6.0
email-validator==2.0.0

# Background Tasks & Scheduling
schedule==1.2.0
celery==5.3.4
redis==5.0.1

# Security
cryptography==41.0.4

# Production Server (optional)
gunicorn==21.2.0

# Development & Monitoring (optional)
flower==2.0.1

# AWS (for potential future SES integration - currently unused)
boto3==1.28.85
botocore==1.31.85